feat: Use ListItem to render suggestion item
