+++ src/drive/web/modules/search/components/styles.styl
@@ -43,39 +43,6 @@
         b
             font-weight bolder
             color var(--scienceBlue)
-    .coz-searchbar-autosuggest-suggestion
-        padding .8em .5em
-        border-width 0
-        border-bottom-width 1px
-        border-style solid
-        border-color var(--silver)
-        cursor pointer
-        &:last-child
-            border-bottom-width 0
-    .coz-searchbar-autosuggest-suggestion-item
-        display flex
-        flex-direction row
-        justify-content flex-start
-        align-items center
-    .coz-searchbar-autosuggest-suggestion-icon
-        max-height 2rem
-        min-height 2rem
-        margin-right 1rem
-    .coz-searchbar-autosuggest-suggestion-content
-        flex-grow 1
-        white-space nowrap
-        overflow hidden
-        text-overflow ellipsis
-        display flex
-        flex-direction column
-    .coz-searchbar-autosuggest-suggestion-subtitle
-        color var(--coolGrey)
-    .coz-searchbar-autosuggest-suggestion-highlighted
-        background var(--paleGrey)
-        box-shadow inset 4px 0 0 0 var(--dodgerBlue)
-    .coz-searchbar-autosuggest-section-title
-        padding .5em
-        font-size .8em
 
     .bar-search-container
         height 100%
